refactor(MetadataFile): moving files around
